python-monitor/requirements.txt
telethon>=1.43.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
google-genai>=1.0.0
tenacity>=8.2.0
redis[hiredis]>=5.0.0
